Moving product specific code out of analytics api and clients
