+++ gdrive/sheets_client.py
@@ -23,8 +23,6 @@ quite a bit.
 Some batching in the future if the use case for this library gets heavier is necessary. 
 """
 
-# Generic functions
-
 
 def update_cell_value(
     sheet_id: str, page_name: str, range_str: str, value: str, vio="USER_ENTERED"
@@ -180,406 +178,6 @@ def export_df_to_gdrive_speadsheet(df: pd.DataFrame, sheets_id: str, title="Shee
         raise error.ExportError(result["error"]["message"])
 
     return result
-
-
-# Project specific functions
-
-
-def create_pivot_tables(
-    df: pd.DataFrame, page_names: (str, str, str), names_to_id: dict, sheets_id: str
-):
-    # Make a dictionary mapping the name of the column to its index, useful for the pivot tables.
-    col_dict = {}
-    for idx, val in enumerate(df.iloc[0]):
-        col_dict[val] = idx
-
-    create_first_visit_pt(sheets_id, names_to_id[page_names[0]], col_dict)
-    log.info(
-        "Added 2 pivot tables to %s (%s)" % (page_names[0], names_to_id[page_names[0]])
-    )
-
-    create_session_start_pt(sheets_id, names_to_id[page_names[1]], col_dict)
-    log.info(
-        "Added 2 pivot tables to %s (%s)" % (page_names[1], names_to_id[page_names[1]])
-    )
-
-    create_clicks_pt(sheets_id, names_to_id[page_names[2]], col_dict)
-    log.info(
-        "Added pivot table to %s (%s)" % (page_names[2], names_to_id[page_names[2]])
-    )
-
-    create_feedback_pt(sheets_id, names_to_id[page_names[3]], col_dict)
-    log.info(
-        "Added pivot table to %s (%s)" % (page_names[3], names_to_id[page_names[3]])
-    )
-
-    update_cell_value(sheets_id, page_names[0], "A17", "Total First Visits")
-    update_cell_value(
-        sheets_id,
-        page_names[0],
-        "A18",
-        '=GETPIVOTDATA("SUM of eventCount",A1, "eventName", "first_visit") + GETPIVOTDATA("SUM of eventCount",F1, "eventName", "first_visit")',
-    )
-    log.info("Wrote totals to %s" % (page_names[0]))
-
-    update_cell_value(sheets_id, page_names[1], "A17", "Total Sessions")
-    update_cell_value(
-        sheets_id,
-        page_names[1],
-        "A18",
-        '=GETPIVOTDATA("SUM of eventCount",A1, "eventName", "session_start") + GETPIVOTDATA("SUM of eventCount",F1, "eventName", "session_start")',
-    )
-    log.info("Wrote totals to %s" % (page_names[1]))
-
-
-def create_first_visit_pt(sheets_id, page_id, col_dict):
-    first_visit_facebook_pt_def = {
-        "pivotTable": {
-            "source": {
-                # First Sheet (Sheet1) is always ID 0
-                "sheetId": 0,
-            },
-            "rows": [
-                {
-                    "sourceColumnOffset": col_dict["eventName"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["firstUserSource"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["eventCount"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-            ],
-            "filterSpecs": [
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "first_visit",
-                                }
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["eventName"],
-                },
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "facebook",
-                                },
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["firstUserSource"],
-                },
-            ],
-            "values": [
-                {
-                    "summarizeFunction": "SUM",
-                    "sourceColumnOffset": col_dict["eventCount"],
-                }
-            ],
-            "valueLayout": "HORIZONTAL",
-        }
-    }
-    first_visit_rt_pt_def = {
-        "pivotTable": {
-            "source": {
-                # First Sheet (Sheet1) is always ID 0
-                "sheetId": 0,
-            },
-            "rows": [
-                {
-                    "sourceColumnOffset": col_dict["eventName"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["firstUserSource"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["eventCount"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-            ],
-            "filterSpecs": [
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "first_visit",
-                                }
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["eventName"],
-                },
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "rt",
-                                },
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["firstUserSource"],
-                },
-            ],
-            "values": [
-                {
-                    "summarizeFunction": "SUM",
-                    "sourceColumnOffset": col_dict["eventCount"],
-                }
-            ],
-            "valueLayout": "HORIZONTAL",
-        }
-    }
-
-    add_pivot_tables(
-        sheets_id,
-        page_id,
-        first_visit_facebook_pt_def,
-    )
-    add_pivot_tables(
-        sheets_id,
-        page_id,
-        first_visit_rt_pt_def,
-        row_idx=0,
-        col_idx=5,
-    )
-
-
-def create_session_start_pt(sheets_id, page_id, col_dict):
-    # Add sessions pivot table, facebook
-    sessions_facebook_pt_def = {
-        "pivotTable": {
-            "source": {
-                # First Sheet (Sheet1) is always ID 0
-                "sheetId": 0,
-            },
-            "rows": [
-                {
-                    "sourceColumnOffset": col_dict["eventName"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["firstUserSource"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["eventCount"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-            ],
-            "filterSpecs": [
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "session_start",
-                                }
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["eventName"],
-                },
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "facebook",
-                                },
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["firstUserSource"],
-                },
-            ],
-            "values": [
-                {
-                    "summarizeFunction": "SUM",
-                    "sourceColumnOffset": col_dict["eventCount"],
-                }
-            ],
-            "valueLayout": "HORIZONTAL",
-        }
-    }
-    sessions_rt_pt_def = {
-        "pivotTable": {
-            "source": {
-                # First Sheet (Sheet1) is always ID 0
-                "sheetId": 0,
-            },
-            "rows": [
-                {
-                    "sourceColumnOffset": col_dict["eventName"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["firstUserSource"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["eventCount"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-            ],
-            "filterSpecs": [
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "session_start",
-                                }
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["eventName"],
-                },
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "rt",
-                                },
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["firstUserSource"],
-                },
-            ],
-            "values": [
-                {
-                    "summarizeFunction": "SUM",
-                    "sourceColumnOffset": col_dict["eventCount"],
-                }
-            ],
-            "valueLayout": "HORIZONTAL",
-        }
-    }
-
-    add_pivot_tables(sheets_id, page_id, sessions_facebook_pt_def)
-    add_pivot_tables(sheets_id, page_id, sessions_rt_pt_def, row_idx=0, col_idx=5)
-
-
-def create_clicks_pt(sheets_id, page_id, col_dict):
-    clicks_pt_def = {
-        "pivotTable": {
-            "source": {
-                # First Sheet (Sheet1) is always ID 0
-                "sheetId": 0,
-            },
-            "rows": [
-                {
-                    "sourceColumnOffset": col_dict["eventName"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["eventCount"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-            ],
-            "values": [
-                {
-                    "summarizeFunction": "SUM",
-                    "sourceColumnOffset": col_dict["eventCount"],
-                }
-            ],
-            "valueLayout": "HORIZONTAL",
-        }
-    }
-
-    add_pivot_tables(sheets_id, page_id, clicks_pt_def)
-
-
-def create_feedback_pt(sheets_id, page_id, col_dict):
-    feedback_pt_def = {
-        "pivotTable": {
-            "source": {
-                "sheetId": 0,
-            },
-            "rows": [
-                {
-                    "sourceColumnOffset": col_dict["eventName"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-                {
-                    "sourceColumnOffset": col_dict["eventCount"],
-                    "showTotals": True,
-                    "sortOrder": "ASCENDING",
-                },
-            ],
-            "filterSpecs": [
-                {
-                    "filterCriteria": {
-                        "condition": {
-                            "type": "TEXT_CONTAINS",
-                            "values": [
-                                {
-                                    "userEnteredValue": "feedback",
-                                }
-                            ],
-                        },
-                        "visibleByDefault": True,
-                    },
-                    "columnOffsetIndex": col_dict["linkUrl"],
-                },
-            ],
-            "values": [
-                {
-                    "summarizeFunction": "SUM",
-                    "sourceColumnOffset": col_dict["eventCount"],
-                }
-            ],
-            "valueLayout": "HORIZONTAL",
-        }
-    }
-
-    add_pivot_tables(sheets_id, page_id, feedback_pt_def)
 
 
 def upload_participant(
